requirements.txt
# Core dependencies
pyaudio>=0.2.11
webrtcvad>=2.0.10
numpy>=1.19.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# File management
watchdog>=2.1.0

# Audio processing
librosa>=0.8.0

# Google Drive integration
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
google-api-python-client>=2.0.0

# Transcription
openai-whisper>=20230124

# System tray UI
pystray>=0.19.0
Pillow>=8.0.0

# Configuration
pyyaml>=6.0
watchdog>=2.1.0